+++ src/LocalData/LocalDataPreEvent.php
@@ -8,11 +8,8 @@ use Symfony\Contracts\EventDispatcher\Event;
 
 class LocalDataPreEvent extends Event
 {
-    private array $options;
-
-    public function __construct(array $options)
+    public function __construct(private array $options)
     {
-        $this->options = $options;
     }
 
     public function getOptions(): array